Overwatch Aim Assistance
A tool that helps a player aim at a target in Overwatch.

The program is written in C++ and uses Windows GDI to capture a
screenshot. The screenshot is scanned in an attempt to find a player's
health bar (pictures in Screenshot folder perfectly demonstrate this).
Upon finding one, the program will scan for the player's border from
which it can obtain the coordinates of the player's head or body.

The program can then either move the cursor smoothly to the desired
location so as to simulate human like movements or just snap directly to
the target location. Trigger bot functionality was also implemented,
however just like all other external colour recognition software, the
crosshair must cross the target's border in order to fire.

Limitations/Room for improvement:
- Shield/Armor (yellow/white) bars were not included in the function
that calculates the width of the health bar. Should be easy to
implement.
- Program only scans for red borders which can occasionally change to
other colours. Can use opencv's canny function to overcome this or scan
for different colours too.
- Mouse movements can be improved in relation to varying sensitivities
and distance to target.
- Currently no threading.